Some basic generation
Each iteration drains the stack as it tries successively smaller n-gram
chains, key for looking up chains is progressively decremented from the
largest n-gram, which only works as long as keys are contiguous. In the
event of a match being found, we take the first one - it would be more
interesting to find matches across all chains and choose a random one.

Generated sentences are short; this may be the result of a lack of
training data, given the m.n-gram approach.

When the stack has been fully drained but desired sentence length has
not been achieved, we should add a full stop, re-seed the stack and
continue.